requirements_no_cuda.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrdict==2.0.1
attrdict3==2.0.2
beautifulsoup4==4.11.2
blosc2==2.0.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chumpy==0.70
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
cython==3.0.11
decorator==4.4.2
distro==1.8.0
dukpy==0.3.0
egi-pynetstation==0.0.0rc1
elementpath==4.7.0
esprima==4.0.1
et-xmlfile==1.1.0
ffpyplayer==4.4.0
filelock==3.14.0
fonttools==4.55.3
freetype-py==2.3.0
fsspec==2024.12.0
future==0.18.3
gevent==22.10.2
gitdb==4.0.10
gitpython==3.1.31
greenlet==2.0.2
html2text==2020.1.16
idna==3.10
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-resources==5.12.0
javascripthon==0.12
jedi==0.18.2
jinja2==3.1.5
jmespath==0.10.0
json-tricks==3.17.3
kiwisolver==1.4.8
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.7.0
mdurl==0.1.2
minio==7.2.15
#mmcv==2.1.0
#mmdet==3.2.0
#mmengine==0.10.5
#mmpose==1.3.2
model-index==0.1.11
mouseinfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
msgpack-numpy==0.4.8
munkres==1.1.4
networkx==3.4.2
ninja==1.11.1.3
ntplib==0.4.0
numexpr==2.8.4
numpy==1.26.4
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.2.3
parso==0.8.3
pillow==11.0.0
pip==25.0.1
platformdirs==4.3.6
polars==1.21.0
proglog==0.1.10
psutil==6.1.1
psychopy==2024.2.5
psychtoolbox==3.0.19.0
py-cpuinfo==9.0.0
pyautogui==0.9.54
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pygetwindow==0.0.9
pygments==2.18.0
pymsgbox==1.0.9
pyparallel==0.2.2
pyparsing==3.2.0
pyperclip==1.9.0
#pypi-search==1.2.1
pyqt5==5.15.9
pyqt5-qt5==5.15.2
pyqt5-sip==12.11.1
pyqt6==6.7.1
pyqt6-qt6==6.7.3
pyqt6-sip==13.8.0
pyrect==0.2.0
pyscreeze==1.0.1
pyserial==3.5
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2023.4
pyyaml==6.0.2
pyzmq==25.0.1
questplus==2023.1
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.4.2
scipy==1.14.1
shapely==2.0.6
six==1.17.0
smmap==5.0.0
soundfile==0.12.1
soupsieve==2.4
sympy==1.13.1
tables==3.8.0
tabulate==0.9.0
termcolor==2.5.0
terminaltables==3.1.10
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.65.2
typing-extensions==4.12.2
tzdata==2024.2
ujson==5.7.0
urllib3==1.26.20
uv==0.6.0
websocket-client==1.5.1
websockets==13.1
wheel==0.45.1
xarray==2023.1.0
xlrd==2.0.1
xmlschema==3.4.3
xtcocotools==1.14.3
yapf==0.43.0
zipp==3.15.0
zope-event==4.6
zope-interface==6.0
